refactor(chat-service): external provider 영역을 helper module 로 분리
# 동기
ChatService.ts (1159 lines / lint 기준 774) 의 max-lines warning 해소
시리즈의 1차 PR. external provider 영역 (3 메서드 / ~300 lines) 이 가장
self-contained 라 첫 split candidate.

# 변경
- backend/api/src/services/chat-service/external-provider.ts (신규, 317 lines)
  - streamFromExternalProvider — multi-turn tool calling 포함
  - executeExternalTool — MCP tool 실행 + tier 권한
  - recordExternalUsageFireAndForget — fire-and-forget usage 기록
  - ExternalProviderDeps interface — ChatService state 를 deps 로 전달
    (providerRouter / currentUserContext / mcpToolResultCallback /
    onUsage / allowedTools). class state 의존성 0, function-pure.
- backend/api/src/services/ChatService.ts (-264 lines)
  - 3 메서드 본문 제거 → helper 위임
  - streamFromExternalProvider 만 thin facade (외부 호출처 1군데)
  - executeExternalTool/recordExternalUsageFireAndForget facade 는 제거
    (external-provider 안에서 직접 호출만 됨)
  - EXTERNAL_LLM_TOOL_BLACKLIST / isPersistableUserId import 제거

# 결과
- ChatService.ts: 774 → 560 lines (lint 기준), -214 line 감소
- 책임 분리 명확화 — chat-service/ 디렉토리의 기존 8 helper module 패턴 답습
- 외부 호출 시그너처 영향 0 (streamFromExternalProvider 시그너처 동일)

# 검증
- tsc 0
- 전체 회귀: 97 suites / 1604 tests / 0 failed (동일)

# 후속 (별도 PR)
- ChatService 의 max-lines warning 잔존 (560 > 400). processMessageInternal
  ~360 lines 의 helper 추출이 다음 candidate.
